Changed stwtypes.h for LLP64 and LP64 compiler data models
Windows 64-bit compilers use the LLP64 data model, Unix compilers the LP64 model.
With the previous declarations in stwtypes.h the type uint32 had a size of 8 bytes
for LP64 compilers.
This patch defines the stwtypes.h in a portable way and fixes casts between
int64 and qint64 types.

+++ opensyde_tool/src/opensyde_gui_elements/spin_box/C_OgeSpxInt64AutoFix.cpp
@@ -242,24 +242,24 @@ void C_OgeSpxInt64AutoFix::m_Init(void)
          if (this->GetIsUnsigned() == true)
          {
             u64_ScaledMin = static_cast<uint64>(f64_Tmp);
-            c_ScaledMin = QVariant(u64_ScaledMin);
+            c_ScaledMin = QVariant(static_cast<quint64>(u64_ScaledMin));
          }
          else
          {
             s64_ScaledMin = static_cast<sint64>(f64_Tmp);
-            c_ScaledMin = QVariant(s64_ScaledMin);
+            c_ScaledMin = QVariant(static_cast<qint64>(s64_ScaledMin));
          }
          this->mc_UnscaledMax.GetAnyValueAsFloat64(f64_Tmp, this->mu32_Index);
          f64_Tmp = C_OSCUtils::h_GetValueScaled(f64_Tmp, this->mf64_Factor, this->mf64_Offset);
          if (this->GetIsUnsigned() == true)
          {
             u64_ScaledMax = static_cast<uint64>(f64_Tmp);
-            c_ScaledMax = QVariant(u64_ScaledMax);
+            c_ScaledMax = QVariant(static_cast<quint64>(u64_ScaledMax));
          }
          else
          {
             s64_ScaledMax = static_cast<sint64>(f64_Tmp);
-            c_ScaledMax = QVariant(s64_ScaledMax);
+            c_ScaledMax = QVariant(static_cast<qint64>(s64_ScaledMax));
          }
       }
       else
@@ -268,15 +268,15 @@ void C_OgeSpxInt64AutoFix::m_Init(void)
          {
             C_SdNdeDpContentUtil::h_GetAnyValueAsUint64(this->mc_UnscaledMin, u64_ScaledMin, this->mu32_Index);
             C_SdNdeDpContentUtil::h_GetAnyValueAsUint64(this->mc_UnscaledMax, u64_ScaledMax, this->mu32_Index);
-            c_ScaledMin = QVariant(u64_ScaledMin);
-            c_ScaledMax = QVariant(u64_ScaledMax);
+            c_ScaledMin = QVariant(static_cast<quint64>(u64_ScaledMin));
+            c_ScaledMax = QVariant(static_cast<quint64>(u64_ScaledMax));
          }
          else
          {
             C_SdNdeDpContentUtil::h_GetAnyValueAsSint64(this->mc_UnscaledMin, s64_ScaledMin, this->mu32_Index);
             C_SdNdeDpContentUtil::h_GetAnyValueAsSint64(this->mc_UnscaledMax, s64_ScaledMax, this->mu32_Index);
-            c_ScaledMin = QVariant(s64_ScaledMin);
-            c_ScaledMax = QVariant(s64_ScaledMax);
+            c_ScaledMin = QVariant(static_cast<qint64>(s64_ScaledMin));
+            c_ScaledMax = QVariant(static_cast<qint64>(s64_ScaledMax));
          }
       }
       this->SetMinimum(c_ScaledMin);
